pyproject.toml
[project]
name = "pet-social-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=4.0.0",
    "aiosqlite>=0.21.0",
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pytest-mock>=3.14.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy[asyncio]>=2.0.40",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
    "types-passlib>=1.7.7.20250401",
    "types-python-jose>=3.4.0.20250224",
]
